Conflicts:
	Controllers/EditSubscriberController.cs
	Controllers/IControllers/INumberController.cs
	Controllers/IControllers/ISearchController.cs
	Controllers/IControllers/ISubscriberController.cs
	Controllers/SearchController.cs
	DatabaseUtils.cs
	View/FormSubscriber.Designer.cs
	View/FormSubscriber.cs

+++ View/FormDetailedInf.cs
@@ -11,16 +11,16 @@ using BillingSystem.Controllers;
 namespace BillingSystem.View
 {
     public partial class FormDetailedInf : Form
-    {
+    {       
         private IDetailedInfController _controler;
-        public FormDetailedInf(int id)
+        public FormDetailedInf(long id)
         {
             //_subscriber = //получить из БД
             InitializeComponent();
         }
 
         private void btnSearch_Click(object sender, EventArgs e)
-        {
+        {            
             _controler.Search(dateTimePickerFrom.Value, dateTimePickerTo.Value, listBoxPhoneNumbers.SelectedItem.ToString());
         }
         // TODO: Сдеать, чтобы при нажатии кнопки "Поиск" контроллер переключался на соответствующий выбронному радиобаттону и затем вызывал метод Search
